close #41
Admin user can now properly delete shippers and carriers, and when they
are deleted, all of their dependent sub-data is deleted (shipments,
bids, and notifications). This is the first major step to resolving any
"ripple effects" that come from deleting accounts.

All account types can now easily be deleted and notifications are sent
out to any users affected by the deletion. Vehicles, contacts, and
drivers are still left over in the database but this should be easy to
fix/clean up.

+++ webrouting/src/main/java/edu/sru/thangiah/webrouting/controller/UserController.java
@@ -21,13 +21,19 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import edu.sru.thangiah.webrouting.domain.Bids;
 import edu.sru.thangiah.webrouting.domain.Carriers;
+import edu.sru.thangiah.webrouting.domain.Notification;
 import edu.sru.thangiah.webrouting.domain.Role;
+import edu.sru.thangiah.webrouting.domain.Shipments;
 import edu.sru.thangiah.webrouting.domain.User;
 import edu.sru.thangiah.webrouting.mailsending.Emailing;
 import edu.sru.thangiah.webrouting.mailsending.MailSending;
+import edu.sru.thangiah.webrouting.repository.BidsRepository;
 import edu.sru.thangiah.webrouting.repository.CarriersRepository;
+import edu.sru.thangiah.webrouting.repository.NotificationRepository;
 import edu.sru.thangiah.webrouting.repository.RoleRepository;
+import edu.sru.thangiah.webrouting.repository.ShipmentsRepository;
 import edu.sru.thangiah.webrouting.repository.UserRepository;
 import edu.sru.thangiah.webrouting.services.SecurityService;
 import edu.sru.thangiah.webrouting.services.UserService;
@@ -58,6 +64,12 @@ public class UserController {
 	
 	private CarriersRepository carriersRepository;
 	
+	private ShipmentsRepository shipmentsRepository;
+	
+	private BidsRepository bidsRepository;
+	
+	private NotificationRepository notficationRepository;
+	
 	@Autowired
     private Emailing emailImpl;
 	
@@ -77,10 +89,13 @@ public class UserController {
 	 * @param roleRepository Used to interact with the roles in the database
 	 * @param carriersRepository Used to interact with the carriers in the database
 	 */
-    public UserController(UserRepository userRepository, RoleRepository roleRepository, CarriersRepository carriersRepository) {
+    public UserController(UserRepository userRepository, RoleRepository roleRepository, CarriersRepository carriersRepository, ShipmentsRepository shipmentsRepository, BidsRepository bidsRepository, NotificationRepository notificationRepository) {
 		this.userRepository = userRepository;
 		this.roleRepository = roleRepository;
 		this.carriersRepository = carriersRepository;
+		this.shipmentsRepository = shipmentsRepository;
+		this.bidsRepository = bidsRepository;
+		this.notficationRepository = notificationRepository;
 	}
     
     /**
@@ -324,14 +339,6 @@ public class UserController {
         User user = userRepository.findById(id)
           .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
         
-        User loggedInUser = getLoggedInUser();
-        if (!user.getShipments().isEmpty()) {
-        	model.addAttribute("error", "Unable to delete due to dependency conflict.");
-        	Logger.error("{} was unable to delete user with ID {} due to depedency conflict.", loggedInUser.getUsername(), user.getId());
-        	model.addAttribute("userstable", userRepository.findAll());
-        	return "users";
-        	
-        }
         model.addAttribute("users", user);
         return "/delete/deleteuserconfirm";
     }
@@ -348,6 +355,34 @@ public class UserController {
           .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
         
         User loggedInUser = getLoggedInUser();
+        
+      if (!user.getShipments().isEmpty()) {
+        	
+        	for(Shipments shipment: user.getShipments()) {
+        		
+        		List<Bids> bids = (List<Bids>) shipment.getBids();
+            	User bidUser;
+            	for (Bids bid : bids) 
+            	{ 
+            		bidUser = CarriersController.getUserFromCarrier(bid.getCarrier());
+            		NotificationController.addNotification(bidUser, "ALERT: Your bid with ID " + bid.getId() + " placed on shipment with ID " + bid.getShipment().getId() + " was deleted because the shipment was deleted");
+            		bidsRepository.delete(bid); 
+            	}
+        		
+        		Logger.info("{} successfully deleted a shipment with ID {}.", user.getUsername(), shipment.getId());
+        		shipmentsRepository.delete(shipment);
+        	}
+        	
+        	model.addAttribute("userstable", userRepository.findAll());
+        	
+        }
+      
+      	List<Notification> notifications = user.getNotifications();
+  		for(Notification n: notifications)
+  		{
+  			notficationRepository.delete(n);
+  		}
+        
         userRepository.delete(user);
       if(user.getRole().toString().equals("SHIPPER")){
     	  return "redirect:/ShipperAdministrationPage";
